refactor:[#168] Answer 도메인 서비스 활용 및 AnswerMetric 복합키 도입
  AnswerDomainService 활용 개선:
  - Application → Domain 서비스로 도메인 로직 위임
  - createAnswer(), validateOwnership(), transitionStatus() 활용
  - QuestionService, UserAccountService 직접 의존 제거

  AnswerMetric 복합키 구조로 변경:
  - AnswerMetricId 복합키 클래스 신규 생성 (@EmbeddedId)
  - Answer-Metric 조합 유일성을 JPA 레벨에서 보장
  - BaseSoftDeleteEntity → BaseTimeEntity 변경 (deleted_at 제거)
  - Repository, Test 코드 복합키 구조에 맞게 수정

  Answer DTO Response 패키지 구조 개선:
  - common, detail, feedback, list, submit 서브패키지로 분리
