+++ src/main/java/com/ktb/answer/controller/AnswerController.java
@@ -8,11 +8,11 @@ import com.ktb.answer.dto.request.AnswerDetailRequest;
 import com.ktb.answer.dto.request.AnswerListRequest;
 import com.ktb.answer.dto.request.AnswerSubmitRequest;
 import com.ktb.answer.dto.request.SessionAnswerSubmitRequest;
-import com.ktb.answer.dto.response.AnswerDetailResponse;
-import com.ktb.answer.dto.response.AnswerListResponse;
-import com.ktb.answer.dto.response.AnswerSubmitResponse;
-import com.ktb.answer.dto.response.FeedbackResponse;
-import com.ktb.answer.dto.response.SessionAnswerSubmitResponse;
+import com.ktb.answer.dto.response.detail.AnswerDetailResponse;
+import com.ktb.answer.dto.response.list.AnswerListResponse;
+import com.ktb.answer.dto.response.submit.AnswerSubmitResponse;
+import com.ktb.answer.dto.response.feedback.FeedbackResponse;
+import com.ktb.answer.dto.response.submit.SessionAnswerSubmitResponse;
 import com.ktb.answer.service.AiFeedbackOrchestrator;
 import com.ktb.answer.service.AnswerApplicationService;
 import com.ktb.auth.security.adapter.SecurityUserAccount;
